go-eicio: Fixed failure to retrieve event at EOF for gzip files
The gzip reader can return successfully and also give EOF error.  This
is now handled properly.  The bug was hidden by the fact that I had
forced gzip to do an extra flush before closing in write routine, which
made the gzip reader not return EOF until the next read.