proto: Added ability to have footer information
In order to accomodate things such as catalogs at the end of files, the
BucketHeader can now be of multiple types, one of which being "FOOTER".
The idea is that the writer can write a footer BucketHeader and then the
payload of the footer is a magic number followed by a fixed-length
number representing the size of the preceding BucketHeader protobuf
message.  A reader can then look at the end of the file and be able to
access the information contained in the footer BucketHeader.