py-proio: rearranged model modules to not make protobuf module angry
protobuf may check to make sure that modules creating descriptors
directly are generated, and it does this by checking that the file name
ends with '_pb2.py'.  I don't want the import path to have this '_pb2',
so I renamed the files, which caused an exception.  Now, the modules are
wrapped in an '__init__.py'.