py-proio/proio/reader.py: fix for latest version of lz4
Latest lz4 release, 0.12.0, changed the return of decompress() from a
byte string to a tuple of a byte string and apparently number of bytes
(documentation doesn't seem to have caught up).  This commit changes
Reader to accommodate both, by first trying to unpack the tuple, and
falling back to a single return value if necessary.