go-proio: fixed a performance issue for lz4
Previously, the (*Writer).Push method directly called the lz4 Writer's
Write method as is done with the gzip writer.  However, this caused a
new lz4 block to be created on every single write.  This caused a
significant performance hit on write, but also a VERY significant
performance hit on read when it encountered many tiny blocks.

The behavior of the Writer has been changed (though it still writes lz4 frames
instead of blocks directly) to create an lz4 Writer only when the bucket
is ready to be written.